Switch the top data columns to match the right order.
Note from Will: "DS18B20 serial number 041682C7F5FF was measuring air
temperature and serial number 041682C8E3FF was measuring the water tank.
So, for the first part of the file, the order is timestamp, air, water.
In the second part of the file, it switches to timestamp, water, air,
notes."